Moved collection package to converter package
