+++ src/main/java/io/beanmapper/BeanMapper.java
@@ -6,12 +6,12 @@ import io.beanmapper.core.BeanField;
 import io.beanmapper.core.BeanFieldMatch;
 import io.beanmapper.core.BeanMatch;
 import io.beanmapper.core.BeanMatchStore;
-import io.beanmapper.core.collections.CollectionListConverter;
-import io.beanmapper.core.collections.CollectionMapConverter;
-import io.beanmapper.core.collections.CollectionSetConverter;
 import io.beanmapper.core.constructor.BeanInitializer;
 import io.beanmapper.core.constructor.NoArgConstructorBeanInitializer;
 import io.beanmapper.core.converter.BeanConverter;
+import io.beanmapper.core.converter.collections.CollectionListConverter;
+import io.beanmapper.core.converter.collections.CollectionMapConverter;
+import io.beanmapper.core.converter.collections.CollectionSetConverter;
 import io.beanmapper.core.converter.impl.*;
 import io.beanmapper.core.unproxy.BeanUnproxy;
 import io.beanmapper.core.unproxy.DefaultBeanUnproxy;
