+++ src/main/java/io/beanmapper/core/BeanMatchStore.java
@@ -4,7 +4,7 @@ import io.beanmapper.annotations.BeanCollection;
 import io.beanmapper.annotations.BeanIgnore;
 import io.beanmapper.annotations.BeanProperty;
 import io.beanmapper.annotations.BeanUnwrap;
-import io.beanmapper.core.collections.BeanCollectionInstructions;
+import io.beanmapper.core.converter.collections.BeanCollectionInstructions;
 import io.beanmapper.core.inspector.PropertyAccessor;
 import io.beanmapper.core.inspector.PropertyAccessors;
 import io.beanmapper.exceptions.BeanMissingPathException;
