+++ src/main/java/io/beanmapper/core/converter/collections/BeanCollectionInstructions.java
@@ -1,4 +1,4 @@
-package io.beanmapper.core.collections;
+package io.beanmapper.core.converter.collections;
 
 import io.beanmapper.annotations.BeanCollectionUsage;
